Add new components and update app layout
This commit introduces new components for a training module, "TrainingForm" for user input and "TrainingTable" for displaying data. They are integrated in the main app replacing the previous layout. Necessary dependencies for these components are also added such as "@emotion/react", "@emotion/styled" and "@mui/material". The '.gitignore' file is updated to exclude '.idea' directory, usually created by JetBrains IDEs. The 'next.config.js' file is updated to configure output and base path and app's color scheme is modified in 'globals.css'. Moreover, an 'appveyor.yml' file is introduced for Windows testing platform, AppVeyor. Also the order of dependencies in 'package.json' was changed for better readability.

These changes implement a new feature for users to input their training data and view it as a table which enhances the user experience. Dependencies were added and updated because they are necessary for building the new training module and they solve the runtime errors which were occurring due to version mismatch. Inclusion of 'appveyor.yml' ensures the Windows compatibility of web application.